+++ DesignerMacros/src/services-crud/create-crud/create-crud.ts
@@ -1,5 +1,6 @@
 // services-crud script (see ~/DesignerMacros/src/services-crud folder in Intent.Modules)
 /// <reference path="../../common/domainHelper.ts" />
+/// <reference path="../../common/crudHelper.ts" />
 /// <reference path="../../common/servicesHelper.ts" />
 /// <reference path="../../services-cqrs-crud/_common/onMapDto.ts" />
 
@@ -32,7 +33,9 @@ namespace servicesCrud {
         createStandardFindAllOperation(service, entity, resultDto);
 
         if (primaryKeys.length > 0) {
-            createStandardUpdateOperation(service, entity, folder);
+            if (!privateSettersOnly){
+                createStandardUpdateOperation(service, entity, folder);
+            }
             createStandardDeleteOperation(service, entity);
 
             const operations = DomainHelper.getCommandOperations(entity);     
@@ -53,6 +56,11 @@ namespace servicesCrud {
             return existing;
         }
 
+        let result = CrudHelper.getOrCreateCrudDto(dtoName, entity, true, CrudConstants.dtoFromEntityMappingId, folder, false);
+        onMapDto(result, folder);
+        return result;
+
+        /*
         let dtoManager = new ElementManager(createElement("DTO", dtoName, folder.id), {
             childSpecialization: "DTO-Field",
         });
@@ -71,7 +79,7 @@ namespace servicesCrud {
         dtoManager.addChildrenFrom(DomainHelper.getAttributesWithMapPath(entity))
 
         dtoManager.collapse();
-        return dtoManager.getElement();
+        return dtoManager.getElement();*/
     }
 
     export function createStandardCreateOperation(service: IElementApi, entity: IElementApi, folder: IElementApi) {
@@ -137,7 +145,7 @@ namespace servicesCrud {
         operationManager.addChild("dto", dtoManager.id);
 
 
-        onMapDto(dtoManager.getElement(), false, `${baseName}Create`);
+        onMapDto(dtoManager.getElement(), folder, false, `${baseName}Create`, true);
         dtoManager.collapse();
         operationManager.collapse();
     }
@@ -227,7 +235,7 @@ namespace servicesCrud {
         operationManager.addChildrenFrom(primaryKeys);
         operationManager.addChild("dto", dtoManager.id);
 
-        onMapDto(dtoManager.getElement(), false, `${baseName}Update`);
+        onMapDto(dtoManager.getElement(), folder, false, `${baseName}Update`);
 
         dtoManager.collapse();
         operationManager.collapse();
@@ -294,7 +302,12 @@ namespace servicesCrud {
         operationManager.addChildrenFrom(primaryKeys);
         operationManager.addChild("dto", dtoManager.id);
 
-        onMapDto(dtoManager.getElement(), false, `${operation.getName()}`);
+        if (DomainHelper.isComplexType(operation.typeReference?.getType())) {
+            var resultDto = CrudHelper.getOrCreateCrudDto(CrudHelper.getName(operationManager.getElement(), operation) + "Result",  operation.typeReference?.getType(), false, CrudConstants.dtoFromEntityMappingId, folder, false);
+            operationManager.setReturnType(resultDto.id, operation.typeReference.isCollection, operation.typeReference.isNullable);
+        }
+  
+        onMapDto(dtoManager.getElement(), folder, false, `${operation.getName()}`);
 
         dtoManager.collapse();
         operationManager.collapse();
@@ -305,7 +318,104 @@ namespace servicesCrud {
         let entityName = entityIsMany ? toPascalCase(pluralize(entity.getName())) : toPascalCase(entity.getName());
         return owningAggregate ? `${toPascalCase(owningAggregate.getName())}${entityName}` : entityName;
     }
+/*
+    function getOrCreateCrudDto(
+        command: MacroApi.Context.IElementApi,
+        mappedElement: MacroApi.Context.IElementApi,
+        autoAddPrimaryKey: boolean,
+        mappingTypeSettingId: string,
+        inbound: boolean = false
+    ): MacroApi.Context.IElementApi {
+        if (mappedElement.typeReference == null) throw new Error("TypeReference is undefined");
 
+        let originalVerb = (command.getName().split(/(?=[A-Z])/))[0];
+        let domainName = mappedElement.typeReference.getType().getName();
+        let baseName = command.getMetadata("baseName")
+            ? `${command.getMetadata("baseName")}${domainName}`
+            : domainName;
+        let dtoName = `${originalVerb}${baseName}`;
+        let dto = getOrCreateDto(dtoName, command.getParent());
+        dto.setMetadata("originalVerb", originalVerb);
+        dto.setMetadata("baseName", baseName);
+
+        //dtoField.typeReference.setType(dto.id);
+        const entityCtor: MacroApi.Context.IElementApi = mappedElement.typeReference.getType()
+        .getChildren("Class Constructor")
+        .sort((a, b) => {
+            // In descending order:
+            return b.getChildren("Parameter").length - a.getChildren("Parameter").length;
+        })[0];
+        if (inbound && entityCtor != null) {
+            dto.setMapping([mappedElement.typeReference.getTypeId(), entityCtor.id], mapToDomainConstructorForDtosSettingId);
+            addDtoFieldsForCtor(autoAddPrimaryKey, entityCtor, dto);
+        } else {
+            dto.setMapping(mappedElement.typeReference.getTypeId(), mappingTypeSettingId);
+            addDtoFields(autoAddPrimaryKey, mappedElement, dto);
+        }
+
+        return dto;
+    }
+
+    function addDtoFieldsForCtor(autoAddPrimaryKey: boolean, ctor: MacroApi.Context.IElementApi, dto: MacroApi.Context.IElementApi) {
+    
+        let childrenToAdd = DomainHelper.getChildrenOfType(ctor, "Parameter").filter(x => x.typeId != null && lookup(x.typeId).specialization !== "Domain Service");
+    
+        childrenToAdd.forEach(e => {
+            if (e.mapPath != null) {
+                if (dto.getChildren("Parameter").some(x => x.getMapping()?.getElement()?.id == e.id)) {
+                    return;
+                }
+            }
+            else if (ctor.getChildren("Parameter").some(x => x.getName().toLowerCase() === e.name.toLowerCase())) {
+                return;
+            }
+    
+            let field = createElement("DTO-Field", toPascalCase(e.name), dto.id);
+            field.setMapping(e.mapPath);
+            if (DomainHelper.isComplexTypeById(e.typeId)){
+                let newDto = getOrCreateCommandCrudDto(dto, field, autoAddPrimaryKey, mapFromDomainMappingSettingId, true );
+                field.typeReference.setType(newDto.id);
+            }else{
+                field.typeReference.setType(e.typeId);
+            }
+            field.typeReference.setIsCollection(e.isCollection);
+            field.typeReference.setIsNullable(e.isNullable);
+        });
+    
+        dto.collapse();
+    }
+    
+    function addDtoFields(autoAddPrimaryKey: boolean, mappedElement: MacroApi.Context.IElementApi, dto: MacroApi.Context.IElementApi) {
+        let dtoUpdated = false;
+        let domainElement = mappedElement
+            .typeReference
+            .getType();
+        let attributesWithMapPaths = getAttributesWithMapPath(domainElement);
+        let isCreateMode = dto.getMetadata("originalVerb")?.toLowerCase()?.startsWith("create") == true;
+        for (var keyName of Object.keys(attributesWithMapPaths)) {
+            let entry = attributesWithMapPaths[keyName];
+            if (isCreateMode && isOwnerForeignKey(entry.name, domainElement)) {
+                continue;
+            }
+            if (dto.getChildren("DTO-Field").some(x => x.getName() == entry.name)) {
+                continue;
+            }
+            let field = createElement("DTO-Field", entry.name, dto.id);
+            field.typeReference.setType(entry.typeId);
+            field.typeReference.setIsNullable(entry.isNullable);
+            field.typeReference.setIsCollection(entry.isCollection);
+            field.setMapping(entry.mapPath);
+            dtoUpdated = true;
+        }
+    
+        if (autoAddPrimaryKey && !isCreateMode) {
+            addPrimaryKeys(dto, domainElement, true);
+        }    
+    
+        if (dtoUpdated) {
+            dto.collapse();
+        }
+    }  */
 }
 
 /**