+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpInvocationStatement.cs
@@ -17,8 +17,14 @@ public class CSharpAccessMemberStatement : CSharpStatement
         Member = memberName;
     }
 
-    private CSharpStatement Reference { get; }
-    private CSharpStatement Member { get; }
+	public CSharpAccessMemberStatement(CSharpStatement expression, CSharpStatement memberName, ICSharpReferenceable referenceable) : base($"{expression.ToString().TrimEnd()}.{memberName}", referenceable)
+	{
+		Reference = expression;
+		Member = memberName;
+	}	
+
+	private CSharpStatement Reference { get; }
+    public CSharpStatement Member { get; }
 
     public CSharpAccessMemberStatement WithSemicolon()
     {
@@ -57,15 +63,13 @@ public class CSharpInvocationStatement : CSharpStatement, IHasCSharpStatements
 {
     private bool _withSemicolon = true;
     private CSharpCodeSeparatorType _defaultArgumentSeparator = CSharpCodeSeparatorType.None;
-    private CSharpClassMethod? _invokedMethod;
-
 
 	public CSharpInvocationStatement(string invokable) : base(invokable)
     {
         Expression = new CSharpStatement(invokable);
     }
 
-    public CSharpInvocationStatement(CSharpStatement expression, string member) : base($"{expression.ToString().TrimEnd()}.{member}")
+	public CSharpInvocationStatement(CSharpStatement expression, string member) : base($"{expression.ToString().TrimEnd()}.{member}")
     {
         Expression = new CSharpAccessMemberStatement(expression, member);
     }
@@ -110,71 +114,21 @@ public class CSharpInvocationStatement : CSharpStatement, IHasCSharpStatements
         _withSemicolon = false;
         return this;
     }
-	internal CSharpInvocationStatement Invokes(CSharpClassMethod method)
-	{
-        _invokedMethod = method;
-        return this;
+
+    public bool IsAsyncInvocation()
+    {
+        return Expression.Reference is ICSharpMethodDeclaration method && (method.IsAsync || method.ReturnType?.GetText("").Contains("Task") == true);
 	}
 
 	public override string GetText(string indentation)
     {
-        if (_invokedMethod?.IsAsync == true)
-        {
-            return AsyncAwareInvocationText(indentation);
-        }
         return $"{RelativeIndentation}{Expression.GetText(indentation)}({GetArgumentsText(indentation)}){(_withSemicolon ? ";" : string.Empty)}";
     }
 
-    private string AsyncAwareInvocationText(string indentation)
-    {
-        bool parentIsAsync = true;
-        var parent = GetParentMethodorConstructor();
-        string parentCancellationTokenName = null;
-        if (parent != null)
-        {
-            switch (parent)
-            {
-                case CSharpConstructor ctor:
-                    parentIsAsync = false;
-                    break;
-                case CSharpClassMethod method:
-                    parentIsAsync = method.IsAsync;
-                    var parameter = method.Parameters.FirstOrDefault(p => p.Type is "CancellationToken" or "CancellationToken?");
-					if (parameter != null)
-                    {
-                        parentCancellationTokenName = parameter.Name;
-					}
-                    break;
-			}
-		}
-        if (parentIsAsync)
-        {
-            var argumentCancellationToken = (CSharpStatement)parentCancellationTokenName;
-			argumentCancellationToken.BeforeSeparator = _defaultArgumentSeparator;
-			argumentCancellationToken.AfterSeparator = CSharpCodeSeparatorType.None;
-
-			return $"{RelativeIndentation}await {Expression.GetText(indentation)}({GetArgumentsText(indentation, argumentCancellationToken)}){(_withSemicolon ? ";" : string.Empty)}";
-		}
-		else
-        {
-			return $"{RelativeIndentation}{Expression.GetText(indentation)}({GetArgumentsText(indentation)}).GetAwaiter().GetResult(){(_withSemicolon ? ";" : string.Empty)}";
-		}
-	}
-
-    private IHasCSharpStatements? GetParentMethodorConstructor()
-    {
-        var current = Parent;
-        while (current != null && current is not CSharpClassMethod or CSharpConstructor)
-        {
-            current = (current as CSharpStatement)?.Parent;
-        }
-        return current;
-	}
-
-	private string GetArgumentsText(string indentation, params CSharpStatement[] additionalArguments)
+	private string GetArgumentsText(string indentation)
     {
         var additionalIndentation = GetAdditionalIndentationIfArgsOnNewLines();
-		return Statements.Concat(additionalArguments ?? Enumerable.Empty<CSharpStatement>()).JoinCode(",", $"{indentation}{additionalIndentation}");
+		return Statements.JoinCode(",", $"{indentation}{additionalIndentation}");
     }
 
     private string GetAdditionalIndentationIfArgsOnNewLines()